Moved api to ES6 modules for import/export REQUIRES type: module in package.json Reduces code in app.js
